pyproject.toml
[tool.poetry]
name = "skymap"
version = "0.1.0"
description = ""
license = "GPL-3.0-or-later"
authors = ["Roel Zinkstok <rzinkstok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.24.3"
mysql-connector = "^2.2.9"
deap = "^1.3.3"
astropy = "^5.3"
matplotlib = "^3.7.1"
beautifulsoup4 = "^4.12.2"
astroquery = "^0.4.6"
pillow = "^9.5.0"
jinja2 = "^3.1.2"
rtree = "^1.0.1"
prospector = "^1.10.2"
black = "^23.3.0"
scipy = "^1.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
